organize module structure
