+++ docs/api.rst
@@ -8,7 +8,8 @@ Import snapatac2 as::
     import snapatac2 as snap
 
 Data IO
-~~~~~~~
+-------
+
 (See more at `anndata-docs <https://anndata.readthedocs.io/en/latest/anndata.AnnData.html>`_)
 
 .. autosummary::
@@ -18,6 +19,9 @@ Data IO
     read_h5ad
 
 Preprocessing: `pp`
+-------------------
+
+Basic Preprocessing
 ~~~~~~~~~~~~~~~~~~~
 
 .. autosummary::
@@ -25,3 +29,33 @@ Preprocessing: `pp`
 
     pp.make_tile_matrix
     pp.select_features
+    pp.knn
+
+Batch effect correction
+~~~~~~~~~~~~~~~~~~~~~~~
+
+Tools: `tl`
+-----------
+
+Any transformation of the data matrix that is not *preprocessing*.
+In contrast to a *preprocessing* function, a *tool* usually adds an easily
+interpretable annotation to the data matrix, which can then be visualized with
+a corresponding plotting function.
+
+Embeddings
+~~~~~~~~~~
+
+.. autosummary::
+   :toctree: _autosummary
+
+   tl.spectral
+   tl.umap
+
+Clustering
+~~~~~~~~~~
+
+.. autosummary::
+   :toctree: _autosummary
+
+   tl.kmeans
+   tl.leiden
