--- snapatac2-python/snapatac2/cluster/__init__.py
@@ -1,2 +1,0 @@
-from .leiden import knn, leiden
-from .kmeans import kmeans
