--- snapatac2-python/snapatac2/embedding/__init__.py
@@ -1,1 +1,0 @@
-from ._spectral import spectral
