+++ snapatac2-python/snapatac2/genome.py
@@ -1,0 +1,52 @@
+hg38 = {
+    "chr1": 248956422,
+    "chr2": 242193529,
+    "chr3": 198295559,
+    "chr4": 190214555,
+    "chr5": 181538259,
+    "chr6": 170805979,
+    "chr7": 159345973,
+    "chr8": 145138636,
+    "chr9": 138394717,
+    "chr10": 133797422,
+    "chr11": 135086622,
+    "chr12": 133275309,
+    "chr13": 114364328,
+    "chr14": 107043718,
+    "chr15": 101991189,
+    "chr16": 90338345,
+    "chr17": 83257441,
+    "chr18": 80373285,
+    "chr19": 58617616,
+    "chr20": 64444167,
+    "chr21": 46709983,
+    "chr22": 50818468,
+    "chrX": 156040895,
+    "chrY": 57227415
+}
+
+GRCm39 = {
+    "chr1": 195154279,
+    "chr2": 181755017,
+    "chr3": 159745316,
+    "4": 156860686,
+    "5": 151758149,
+    "6": 149588044,
+    "7": 144995196,
+    "8": 130127694,
+    "9": 124359700,
+    "10": 130530862,
+    "11": 121973369,
+    "12": 120092757,
+    "13": 120883175,
+    "14": 125139656,
+    "15": 104073951,
+    "16": 98008968,
+    "17": 95294699,
+    "18": 90720763,
+    "19": 61420004,
+    "X": 169476592,
+    "Y": 91455967,
+}
+
+
