+++ snapatac2-python/snapatac2/tools/_kmeans.py
@@ -1,66 +1,17 @@
 from typing import Optional, Union, Type
 import pandas as pd
-import scipy.sparse as ss
 import numpy as np
-import leidenalg as la
-from sklearn.neighbors import kneighbors_graph
 from anndata.experimental import AnnCollection
-import igraph as ig
 import anndata as ad
-from leidenalg.VertexPartition import MutableVertexPartition
-from .. import _utils
+import snapatac2._snapatac2 as _snapatac2
 
-def knn(
+def kmeans(
     adata: Union[ad.AnnData, AnnCollection],
-    n_neighbors: int = 15,
-    n_pcs: Optional[int] = None,
-    use_rep: Optional[str] = None,
-    n_jobs: int = -1,
-) -> None:
-    """
-    Compute a neighborhood graph of observations.
-
-    Parameters
-    ----------
-    adata
-        Annotated data matrix.
-    n_neighbors
-        The number of nearest neighbors to be searched.
-    key_added
-        If not specified, connectivities .obsp['connectivities'].
-        connectivities in .obsp[key_added+'_connectivities'].
-
-    Returns
-    -------
-    Depending on `copy`, updates or returns `adata` with the following:
-    See `key_added` parameter description for the storage path of
-    connectivities and distances.
-    **connectivities** : sparse matrix of dtype `float32`.
-        Weighted adjacency matrix of the neighborhood graph of data
-        points. Weights should be interpreted as connectivities.
-    **distances** : sparse matrix of dtype `float32`.
-        Instead of decaying weights, this stores distances for each pair of
-        neighbors.
-    """
-    if use_rep is None:
-        data = adata.obsm["X_spectral"]
-    else:
-        data = adata.obsm[use_rep]
-
-    adj = kneighbors_graph(data, n_neighbors, mode='distance', n_jobs=n_jobs)
-    np.reciprocal(adj.data, out=adj.data)
-    adata.obsp['connectivities'] = adj
-
-# TODO: leiden function in igraph C library is much faster
-def leiden(
-    adata: Union[ad.AnnData, AnnCollection],
-    resolution: float = 1,
-    partition_type: MutableVertexPartition = la.RBConfigurationVertexPartition,
-    use_weights: bool = True,
+    n_clusters: int,
     n_iterations: int = -1,
     random_state: int = 0,
-    key_added: str = 'leiden',
-    adjacency: Optional[ss.spmatrix] = None,
+    use_rep: Optional[str] = None,
+    key_added: str = 'kmeans',
 ):
     """
     Cluster cells into subgroups [Traag18]_.
@@ -79,7 +30,7 @@ def leiden(
         A parameter value controlling the coarseness of the clustering.
         Higher values lead to more clusters.
         Set to `None` if overriding `partition_type`
-        to one that doesn’t accept a `resolution_parameter`.
+        to one that doesn't accept a `resolution_parameter`.
     random_state
         Change the initialization of the optimization.
     restrict_to
@@ -128,27 +79,20 @@ def leiden(
         and `n_iterations`.
     """
 
-    if adjacency is None:
-        adjacency = adata.obsp["connectivities"]
-    gr = _utils.get_igraph_from_adjacency(adjacency)
-    if use_weights:
-        weights = gr.es["weight"]
+    if use_rep is None:
+        data = adata.obsm["X_spectral"]
     else:
-        weights = None
+        data = adata.obsm[use_rep]
 
-    partition = la.find_partition(gr, partition_type, n_iterations=n_iterations,
-        seed=random_state, resolution_parameter=resolution, weights=weights
-    )
-    groups = np.array(partition.membership)
-
+    groups = _snapatac2.kmeans(n_clusters, data)
     adata.obs[key_added] = pd.Categorical(
         values=groups.astype('U'),
         categories=sorted(map(str, np.unique(groups))),
     )
     # store information on the clustering parameters
-    adata.uns['leiden'] = {}
-    adata.uns['leiden']['params'] = dict(
-        resolution=resolution,
+    adata.uns['kmeans'] = {}
+    adata.uns['kmeans']['params'] = dict(
+        n_clusters=n_clusters,
         random_state=random_state,
         n_iterations=n_iterations,
     )
